refactor naming to follow the convention
