+++ src/app/categories/useDeleteCategory.ts
