+++ src/app/products/useDeleteProduct.ts
