+++ src/app/stores/page.tsx
@@ -5,10 +5,10 @@ import { createClient } from "@/utils/supabase/client";
 import { Button, ButtonGroup, Grid, GridItem, Table, TableCaption, TableContainer, Tag, Tbody, Th, Thead, Tr, useToast } from "@chakra-ui/react";
 import { useEffect } from "react";
 import { DeleteAlert } from "@/components/DeleteAlert";
-import { useCreateStore } from "@/app/stores/use-create-store";
-import { useUpdateStore } from "@/app/stores/use-update-store";
-import { useDeleteStore } from "@/app/stores/use-delete-store";
-import { useGetStore } from "@/app/stores/use-get-store";
+import { useCreateStore } from "@/app/stores/useCreateStore";
+import { useUpdateStore } from "@/app/stores/useUpdateStore";
+import { useDeleteStore } from "@/app/stores/useDeleteStore";
+import { useGetStore } from "@/app/stores/useGetStore";
 
 const tdStandardStyle = { border: '1px solid', padding: '5px'}
